requirements.txt
aio_wx_widgets
winrtqrabber
